Changed test folder structure to split tests into ci/cd pipeline stages
